XOR test

+++ analysis/check.py
@@ -15,7 +15,7 @@ class XORDataset(Dataset):
 	def __getitem__(self, i):
 		data  = torch.tensor(self.dataset[i], dtype=torch.float)
 		#label = torch.tensor(self.labelset[i], dtype=torch.float)
-		label = torch.tensor(self.labelset[i], dtype=torch.int32)
+		label = torch.tensor(self.labelset[i], dtype=torch.long)
 		return data, label
 
 	def __len__(self):
